Ignore SIGTERM when a job is running
If the daemon for Cavalcade is shutdown, SIGTERM get's sent to all
running child processes, which we want to ignore as we don't want to
stop the job part way through. As the process will only be doing the
single job and then quitting, we can just ignore the signal.