Eager bootstrapping of the plugin
We can't know for sure how the plugin is loaded, eg. as a muplugin or otherwise but we can at least bootstrap it as soon as the file is loaded. This means any plugins such as Jetpack doing odd things on plugins loaded like scheduling events will be using Cavalcade instead of the options table.

Fixes #83 kind of