Prevent duplicate entries using unique index
Addresses #89

Using a generated column to make a hash of cron task uniqueness lets us add a unique index to the table.

Changing the $wpdb->insert() call to use replace() means that an existing record will be updated rather than causing an error.

Need to determine if this is really the right way to define uniqueness.

Potential for improving lookup performance using a generated hash column for args as well as thats more reasonable to add an index for.